fix: remove print_v call in BaseTypes.__getattr__
This is a fix for something I noticed happening when neuromllite
objects are used in IPython (or PyCharm) interactivelly. It seems that
IPython under the hood calls getattr on objects to inspect them.

See this example executed in IPython:

IPython 7.22.0 -- An enhanced Interactive Python. Type '?' for help.

In [1]: from neuromllite import Base

In [2]: Base()
Out[2]: neuromllite >>> No field or child:
_ipython_canary_method_should_not_exist_ in None
neuromllite >>> No field or child: _repr_mimebundle_ in None
Base (None)

In [3]: quit()

Since IPython is pretty common it would probably be a good idea
to not print from __getattr__ (or have any other side effects
really).